Implement loadRequiredConfigIntoEntity
